command updated. login implemented

+++ app/Console/Commands/InstallAssessmentSuite.php
@@ -14,6 +14,7 @@ namespace App\Console\Commands;
 
 use Dotenv\Dotenv;
 use Illuminate\Console\Command;
+use App\User;
 
 class InstallAssessmentSuite extends Command
 {
@@ -46,6 +47,15 @@ class InstallAssessmentSuite extends Command
      */
     public function handle()
     {
+        $user = User::where('usertype', 'superadmin')->first();
+        if ($user) {
+            $this->error('It seems admin account is already created');
+            if (!$this->confirm('Do you still want to continue?')) {
+                $this->line('Flight Aborted. Bye!');
+
+                return;
+            }
+        }
         if (!$this->confirm('Do you want to install Assessment Suite?')) {
             $this->line('Flight Aborted. Bye!');
 
@@ -59,7 +69,7 @@ class InstallAssessmentSuite extends Command
         $this->configureDrivers();
         $this->makeMigrations();
         $this->seedDatabase();
-
+        $this->createSuperAdmin();
         $this->info('Assessment Suite is installed ⚡');
     }
 
@@ -363,5 +373,30 @@ class InstallAssessmentSuite extends Command
         $this->call('db:seed');
         $this->info('Database Seeded.');
         $this->info('--------------------------------------------------------------');
+    }
+
+    public function createSuperAdmin()
+    {
+        $config['name'] = null;
+        $config['email'] = null;
+        $config['password'] = null;
+
+        $config['name'] = $this->ask('Name of administrator? ', $config['name']);
+
+        $config['email'] = $this->ask('Email address of administrator?', $config['email']);
+
+        $config['password'] = $this->secret('Password for administrator account? (E.g. secret)', $config['password']);
+
+        $this->formatConfigsTable($config);
+
+        if (!$this->confirm('Are these credentials correct?')) {
+            return $this->createSuperAdmin();
+        }
+        $user = new User();
+        $user->name = $config['name'];
+        $user->email = $config['email'];
+        $user->password = bcrypt($config['password']);
+        $user->save();
+        $this->info('Administrator account created successfully');
     }
 }
